+++ src/index.jsx
@@ -1,34 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import Page from './components/Page';
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function addByOne() {
-    setState({
-      count: count + 1,
-    });
-  }
-
-  function addByClickedNumber(clickedNumber) {
-    setState({
-      count: count + clickedNumber,
-    });
-  }
-
-  return (
-    <Page
-      count={count}
-      addByOne={addByOne}
-      addByClickedNumber={addByClickedNumber}
-    />
-  );
-}
+import App from './components/App';
 
 ReactDOM.render(
   <App />,
